pixsim7_backend/requirements.txt
# PixSim7 Backend Dependencies

# ===== Core Framework =====
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ===== Database =====
sqlmodel==0.0.14
sqlalchemy==2.0.25
asyncpg==0.29.0  # Async PostgreSQL driver
alembic==1.13.1  # Database migrations
psycopg2-binary==2.9.9  # Sync PostgreSQL driver (for Alembic)
pgvector==0.2.4  # PostgreSQL vector extension for similarity search

# ===== Cache & Queue =====
redis==5.0.1
arq==0.25.0  # Async task queue

# ===== HTTP Client =====
httpx==0.26.0  # Async HTTP client
aiohttp==3.9.1

# ===== Video Generation Providers =====
# Install from local path or GitHub
# Local: pip install -e G:/code/pixverse-py
# GitHub: pip install git+https://github.com/user/pixverse-py.git
pixverse-py  # Pixverse AI SDK

# Local: pip install -e G:/code/sora-py
# GitHub: pip install git+https://github.com/user/sora-py.git
sora-py  # OpenAI Sora SDK

# ===== Auth & Security =====
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4  # Password hashing
python-multipart==0.0.6  # Form data parsing

# ===== File Handling =====
pillow==10.2.0  # Image processing
imagehash==4.3.1  # Perceptual hashing

# ===== Utilities =====
python-dotenv==1.0.0  # .env file support
pydantic-extra-types==2.3.0  # Extra Pydantic types

# ===== Development =====
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1  # Code formatting
ruff==0.1.15  # Linting

# ===== Optional: Video Processing (Phase 2) =====
# opencv-python==4.9.0.80
# ffmpeg-python==0.2.0

# ===== System Monitoring =====
psutil==5.9.6  # System metrics (CPU, memory, disk)
